backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
python-dotenv>=1.0.0
langgraph
langchain-core
groq>=0.25.0
# WebSocket support for live streaming
websockets>=12.0
python-multipart>=0.0.6